Improve the code structure - Move wrap methods to CodeWrapper
